cleanup: split native runtime transaction tests
